requirements.txt
# Web framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Auth & security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# passlib 1.7.4 has compatibility issues with bcrypt 4.x on some setups
bcrypt==3.2.2
python-multipart==0.0.9

# Validation & config
pydantic[email]==2.6.1
pydantic-settings==2.1.0

# Utils
python-dotenv==1.0.1
